refactor: rename interfaces to entities and use classes
