+++ src/album/entities/album.entity.ts
@@ -1,4 +1,4 @@
-export interface Album {
+export class Album {
   id: string; // uuid v4
   name: string;
   year: number;
