+++ src/artist/entities/artist.entity.ts
@@ -1,4 +1,4 @@
-export interface Artist {
+export class Artist {
   id: string; // uuid v4
   name: string;
   grammy: boolean;
