+++ src/favorites/entities/favorites.entity.ts
@@ -1,4 +1,4 @@
-export interface Favorites {
+export class Favorites {
   artists: string[]; // favorite artists ids
   albums: string[]; // favorite albums ids
   tracks: string[]; // favorite tracks ids
