--- src/favorites/interfaces/favorites-response.interface.ts
@@ -1,9 +1,0 @@
-import { Artist } from '../../artist/interfaces/artist.interface';
-import { Album } from '../../album/interfaces/album.interface';
-import { Track } from '../../track/interfaces/track.interface';
-
-export interface FavoritesResponse {
-  artists: Artist[];
-  albums: Album[];
-  tracks: Track[];
-}
